Remove obsolete (int) casts on Enumeration comparisons in LINQ statements (See Case 836)
git-svn-id: svn+ssh://dasz.at/srv/dasz/svn/Kistl@1300 76cbe2a5-d633-47ad-8992-a0d0e66ffbf2

+++ Kistl.Server/Generators/EntityFramework/Implementation/EfModel/Model.csdl.cs
@@ -62,7 +62,7 @@ namespace Kistl.Server.Generators.EntityFramework.Implementation.EfModel
         internal static IEnumerable<Relation> GetRelationsWithSeparateStorage(IKistlContext ctx)
         {
             return ctx.GetQuery<Relation>()
-                .Where(r => (int)r.Storage == (int)StorageType.Separate)
+                .Where(r => r.Storage == StorageType.Separate)
                 .ToList()
                 .OrderBy(r => r.GetAssociationName());
         }
@@ -70,7 +70,7 @@ namespace Kistl.Server.Generators.EntityFramework.Implementation.EfModel
         internal static IEnumerable<Relation> GetRelationsWithoutSeparateStorage(IKistlContext ctx)
         {
             return ctx.GetQuery<Relation>()
-                .Where(r => (int)r.Storage != (int)StorageType.Separate)
+                .Where(r => r.Storage != StorageType.Separate)
                 .ToList()
                 .OrderBy(r => r.GetAssociationName());
         }
